--- Samples/BetaSamples/LocalMediaSamples/ComplianceRecordingBot/FrontEnd/CRFrontEnd.csproj
@@ -1,74 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <Import Project="$(MSBuildThisFileDirectory)..\..\..\..\Graph.props" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{739D09C4-47E8-42B7-9B89-94DCB890AC0F}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Sample.ComplianceRecordingBot.FrontEnd</RootNamespace>
-    <AssemblyName>Sample.ComplianceRecordingBot.FrontEnd</AssemblyName>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <NuGetPackageImportStamp>
-    </NuGetPackageImportStamp>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x64'">
-    <DebugSymbols>true</DebugSymbols>
-    <OutputPath>bin\x64\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <DebugType>full</DebugType>
-    <PlatformTarget>x64</PlatformTarget>
-    <LangVersion>latest</LangVersion>
-    <ErrorReport>prompt</ErrorReport>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|x64'">
-    <OutputPath>bin\x64\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <Optimize>true</Optimize>
-    <DebugType>pdbonly</DebugType>
-    <PlatformTarget>x64</PlatformTarget>
-    <ErrorReport>prompt</ErrorReport>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="Bot\Bot.cs" />
-    <Compile Include="Bot\BotMediaStream.cs" />
-    <Compile Include="Bot\CallHandler.cs" />
-    <Compile Include="Bot\ExceptionExtensions.cs" />
-    <Compile Include="Http\ExceptionLogger.cs" />
-    <Compile Include="SampleConstants.cs" />
-    <Compile Include="Http\Controllers\DemoController.cs" />
-    <Compile Include="Http\Controllers\HttpRouteConstants.cs" />
-    <Compile Include="Http\Controllers\PlatformCallController.cs" />
-    <Compile Include="Http\HttpConfigurationInitializer.cs" />
-    <Compile Include="Http\LoggingMessageHandler.cs" />
-    <Compile Include="IConfiguration.cs" />
-    <Compile Include="LRUCache.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Service.cs" />
-    <Compile Include="Utilities.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <Reference Include="Microsoft.Skype.Internal.Media.H264, Version=0.0.0.0, Culture=neutral, processorArchitecture=AMD64">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>..\..\..\..\packages\Microsoft.Skype.Bots.Media\1.14.1.234-alpha\src\skype_media_lib\Microsoft.Skype.Internal.Media.H264.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNet.WebApi.Owin" Version="5.2.7" />
-    <PackageReference Include="Microsoft.Graph.Communications.Calls.Media" Version="1.2.0-beta.1" />
-    <PackageReference Include="Microsoft.Owin.Hosting" Version="4.0.1" />
-    <PackageReference Include="Microsoft.Skype.Bots.Media" Version="1.14.1.234-alpha" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <Reference Include="System.Data.Entity" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\..\Common\Sample.Common.Beta\Sample.Common.Beta.csproj">
-      <Project>{88b67fbe-6de8-4dc3-83cf-9f6b4ea10138}</Project>
-      <Name>Sample.Common.Beta</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-</Project>
