--- Samples/BetaSamples/LocalMediaSamples/ComplianceRecordingBot/FrontEnd/SampleConstants.cs
@@ -1,23 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="SampleConstants.cs" company="Microsoft Corporation">
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license.
-// </copyright>
-// <summary>
-//   The constants.
-// </summary>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace Sample.ComplianceRecordingBot.FrontEnd
-{
-    /// <summary>
-    /// Contants used by the bot.
-    /// </summary>
-    public static class SampleConstants
-    {
-        /// <summary>
-        /// Number of sockets to receive video only.
-        /// </summary>
-        public const uint NumberOfMultiviewSockets = 3;
-    }
-}
