+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/FrontEnd/Bot/Bot.cs
@@ -1,0 +1,266 @@
+﻿// <copyright file="Bot.cs" company="Microsoft Corporation">
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+// </copyright>
+
+namespace Sample.PolicyRecordingBot.FrontEnd.Bot
+{
+    using System;
+    using System.Collections.Concurrent;
+    using System.Collections.Generic;
+    using System.Data;
+    using System.Threading.Tasks;
+    using Microsoft.Graph;
+    using Microsoft.Graph.Communications.Calls;
+    using Microsoft.Graph.Communications.Calls.Media;
+    using Microsoft.Graph.Communications.Client;
+    using Microsoft.Graph.Communications.Common;
+    using Microsoft.Graph.Communications.Common.Telemetry;
+    using Microsoft.Graph.Communications.Resources;
+    using Microsoft.Skype.Bots.Media;
+    using Sample.Common;
+    using Sample.Common.Authentication;
+    using Sample.Common.Logging;
+    using Sample.PolicyRecordingBot.FrontEnd;
+
+    /// <summary>
+    /// The core bot logic.
+    /// </summary>
+    internal class Bot : IDisposable
+    {
+        /// <summary>
+        /// Gets the instance of the bot.
+        /// </summary>
+        public static Bot Instance { get; } = new Bot();
+
+        /// <summary>
+        /// Gets the Graph Logger instance.
+        /// </summary>
+        public IGraphLogger Logger { get; private set; }
+
+        /// <summary>
+        /// Gets the sample log observer.
+        /// </summary>
+        public SampleObserver Observer { get; private set; }
+
+        /// <summary>
+        /// Gets the collection of call handlers.
+        /// </summary>
+        public ConcurrentDictionary<string, CallHandler> CallHandlers { get; } = new ConcurrentDictionary<string, CallHandler>();
+
+        /// <summary>
+        /// Gets the entry point for stateful bot.
+        /// </summary>
+        public ICommunicationsClient Client { get; private set; }
+
+        /// <inheritdoc />
+        public void Dispose()
+        {
+            this.Observer?.Dispose();
+            this.Observer = null;
+            this.Logger = null;
+            this.Client?.Dispose();
+            this.Client = null;
+        }
+
+        /// <summary>
+        /// Initialize the instance.
+        /// </summary>
+        /// <param name="service">Service instance.</param>
+        /// <param name="logger">Graph logger.</param>
+        internal void Initialize(Service service, IGraphLogger logger)
+        {
+            Validator.IsNull(this.Logger, "Multiple initializations are not allowed.");
+
+            this.Logger = logger;
+            this.Observer = new SampleObserver(logger);
+
+            var name = this.GetType().Assembly.GetName().Name;
+            var builder = new CommunicationsClientBuilder(
+                name,
+                service.Configuration.AadAppId,
+                this.Logger);
+
+            var authProvider = new AuthenticationProvider(
+                name,
+                service.Configuration.AadAppId,
+                service.Configuration.AadAppSecret,
+                this.Logger);
+
+            builder.SetAuthenticationProvider(authProvider);
+            builder.SetNotificationUrl(service.Configuration.CallControlBaseUrl);
+            builder.SetMediaPlatformSettings(service.Configuration.MediaPlatformSettings);
+            builder.SetServiceBaseUrl(service.Configuration.PlaceCallEndpointUrl);
+
+            this.Client = builder.Build();
+            this.Client.Calls().OnIncoming += this.CallsOnIncoming;
+            this.Client.Calls().OnUpdated += this.CallsOnUpdated;
+        }
+
+        /// <summary>
+        /// End a particular call.
+        /// </summary>
+        /// <param name="callLegId">
+        /// The call leg id.
+        /// </param>
+        /// <returns>
+        /// The <see cref="Task"/>.
+        /// </returns>
+        internal async Task EndCallByCallLegIdAsync(string callLegId)
+        {
+            try
+            {
+                await this.GetHandlerOrThrow(callLegId).Call.DeleteAsync().ConfigureAwait(false);
+            }
+            catch (Exception)
+            {
+                // Manually remove the call from SDK state.
+                // This will trigger the ICallCollection.OnUpdated event with the removed resource.
+                this.Client.Calls().TryForceRemove(callLegId, out ICall call);
+            }
+        }
+
+        /// <summary>
+        /// Creates the local media session.
+        /// </summary>
+        /// <param name="mediaSessionId">
+        /// The media session identifier.
+        /// This should be a unique value for each call.
+        /// </param>
+        /// <returns>The <see cref="ILocalMediaSession"/>.</returns>
+        private ILocalMediaSession CreateLocalMediaSession(Guid mediaSessionId = default(Guid))
+        {
+            var videoSocketSettings = new List<VideoSocketSettings>();
+
+            // create the receive only sockets settings for the multiview support
+            for (int i = 0; i < SampleConstants.NumberOfMultiviewSockets; i++)
+            {
+                videoSocketSettings.Add(new VideoSocketSettings
+                {
+                    StreamDirections = StreamDirection.Recvonly,
+                    ReceiveColorFormat = VideoColorFormat.H264,
+                });
+            }
+
+            // Create the VBSS socket settings
+            var vbssSocketSettings = new VideoSocketSettings
+            {
+                StreamDirections = StreamDirection.Recvonly,
+                ReceiveColorFormat = VideoColorFormat.H264,
+                MediaType = MediaType.Vbss,
+                SupportedSendVideoFormats = new List<VideoFormat>
+                {
+                    // fps 1.875 is required for h264 in vbss scenario.
+                    VideoFormat.H264_1920x1080_1_875Fps,
+                },
+            };
+
+            // create media session object, this is needed to establish call connections
+            var mediaSession = this.Client.CreateMediaSession(
+                new AudioSocketSettings
+                {
+                    StreamDirections = StreamDirection.Recvonly,
+                    SupportedAudioFormat = AudioFormat.Pcm16K,
+                },
+                videoSocketSettings,
+                vbssSocketSettings,
+                mediaSessionId: mediaSessionId);
+            return mediaSession;
+        }
+
+        /// <summary>
+        /// Incoming call handler.
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="args">The <see cref="CollectionEventArgs{TEntity}"/> instance containing the event data.</param>
+        private void CallsOnIncoming(ICallCollection sender, CollectionEventArgs<ICall> args)
+        {
+            args.AddedResources.ForEach(call =>
+            {
+                // Get the policy recording parameters.
+
+                // The context associated with the incoming call.
+                IncomingContext incomingContext =
+                    call.Resource.IncomingContext;
+
+                // The RP participant.
+                string observedParticipantId =
+                    incomingContext.ObservedParticipantId;
+
+                // If the observed participant is a delegate.
+                IdentitySet onBehalfOfIdentity =
+                    incomingContext.OnBehalfOf;
+
+                // If a transfer occured, the transferor.
+                IdentitySet transferorIdentity =
+                    incomingContext.Transferor;
+
+                string countryCode = null;
+                EndpointType? endpointType = null;
+
+                // Note: this should always be true for CR calls.
+                if (incomingContext.ObservedParticipantId == incomingContext.SourceParticipantId)
+                {
+                    // The dynamic location of the RP.
+                    countryCode = call.Resource.Source.CountryCode;
+
+                    // The type of endpoint being used.
+                    endpointType = call.Resource.Source.EndpointType;
+                }
+
+                IMediaSession mediaSession = Guid.TryParse(call.Id, out Guid callId)
+                    ? this.CreateLocalMediaSession(callId)
+                    : this.CreateLocalMediaSession();
+
+                // Answer call
+                call?.AnswerAsync(mediaSession).ForgetAndLogExceptionAsync(
+                    call.GraphLogger,
+                    $"Answering call {call.Id} with scenario {call.ScenarioId}.");
+            });
+        }
+
+        /// <summary>
+        /// Updated call handler.
+        /// </summary>
+        /// <param name="sender">The <see cref="ICallCollection"/> sender.</param>
+        /// <param name="args">The <see cref="CollectionEventArgs{ICall}"/> instance containing the event data.</param>
+        private void CallsOnUpdated(ICallCollection sender, CollectionEventArgs<ICall> args)
+        {
+            foreach (var call in args.AddedResources)
+            {
+                var callHandler = new CallHandler(call);
+                this.CallHandlers[call.Id] = callHandler;
+            }
+
+            foreach (var call in args.RemovedResources)
+            {
+                if (this.CallHandlers.TryRemove(call.Id, out CallHandler handler))
+                {
+                    handler.Dispose();
+                }
+            }
+        }
+
+        /// <summary>
+        /// The get handler or throw.
+        /// </summary>
+        /// <param name="callLegId">
+        /// The call leg id.
+        /// </param>
+        /// <returns>
+        /// The <see cref="CallHandler"/>.
+        /// </returns>
+        /// <exception cref="ObjectNotFoundException">
+        /// Throws an exception if handler is not found.
+        /// </exception>
+        private CallHandler GetHandlerOrThrow(string callLegId)
+        {
+            if (!this.CallHandlers.TryGetValue(callLegId, out CallHandler handler))
+            {
+                throw new ObjectNotFoundException($"call ({callLegId}) not found");
+            }
+
+            return handler;
+        }
+    }
+}
