+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/FrontEnd/Properties/AssemblyInfo.cs
