+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/Images/TestMeeting1.png
