+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/Images/TestMeeting2.png
