+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/ServiceDefinition.csdef
@@ -1,0 +1,51 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ServiceDefinition name="PolicyRecordingBot" xmlns="http://schemas.microsoft.com/ServiceHosting/2008/10/ServiceDefinition" schemaVersion="2015-04.2.6">
+  <WorkerRole name="CRWorkerRole" vmsize="Standard_D3_v2">
+    <Runtime executionContext="elevated" />
+    <Startup>
+      <Task commandLine="Startup.cmd &gt; Startup.cmd.log" executionContext="elevated" taskType="simple">
+        <Environment>
+          <Variable name="PrivateDefaultCallControlPort">
+            <RoleInstanceValue xpath="/RoleEnvironment/CurrentInstance/Endpoints/Endpoint[@name='DefaultEndpoint']/@port" />
+          </Variable>
+          <Variable name="PrivateInstanceCallControlPort">
+            <RoleInstanceValue xpath="/RoleEnvironment/CurrentInstance/Endpoints/Endpoint[@name='InstanceCallControlEndpoint']/@port" />
+          </Variable>
+          <Variable name="InstanceIpAddress">
+            <RoleInstanceValue xpath="/RoleEnvironment/CurrentInstance/Endpoints/Endpoint[@name='InstanceMediaControlEndpoint']/@address" />
+          </Variable>
+          <Variable name="DefaultCertificate">
+            <RoleInstanceValue xpath="/RoleEnvironment/CurrentInstance/ConfigurationSettings/ConfigurationSetting[@name='DefaultCertificate']/@value" />
+          </Variable>
+        </Environment>
+      </Task>
+    </Startup>
+    <ConfigurationSettings>
+      <Setting name="Microsoft.WindowsAzure.Plugins.Diagnostics.ConnectionString" />
+      <Setting name="ServiceDnsName" />
+      <Setting name="ServiceCNAME" />
+      <!-- CNAME pointing to the .cloudapp.net if available-->
+      <Setting name="DefaultCertificate" />
+      <Setting name="APPINSIGHTS_INSTRUMENTATIONKEY" />
+    </ConfigurationSettings>
+    <Endpoints>
+      <InputEndpoint name="DefaultEndpoint" protocol="tcp" port="443" localPort="9441" />
+      <InstanceInputEndpoint name="InstanceCallControlEndpoint" protocol="tcp" localPort="10100">
+        <AllocatePublicPortFrom>
+          <FixedPortRange max="10199" min="10100" />
+        </AllocatePublicPortFrom>
+      </InstanceInputEndpoint>
+      <InstanceInputEndpoint name="InstanceMediaControlEndpoint" protocol="tcp" localPort="8445">
+        <AllocatePublicPortFrom>
+          <FixedPortRange max="20199" min="20100" />
+        </AllocatePublicPortFrom>
+      </InstanceInputEndpoint>
+    </Endpoints>
+    <LocalResources>
+      <LocalStorage name="NETFXInstall" sizeInMB="1024" cleanOnRoleRecycle="false" />
+    </LocalResources>
+    <Certificates>
+      <Certificate name="Default" storeLocation="LocalMachine" storeName="My" />
+    </Certificates>
+  </WorkerRole>
+</ServiceDefinition>
