+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/WorkerRole/Properties/AssemblyInfo.cs
