+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/WorkerRole/app.config
