+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/WorkerRole/startup.cmd
