+++ Samples/BetaSamples/LocalMediaSamples/PolicyRecordingBot/WorkerRoleContent/diagnostics.wadcfgx
